Fix sudden about:blank's after tab unsuspend.
This bug drove me insane for virtually all the time I'm in
porting this extension to Firefox. Sometimes when I switch
to a tab and then click unsuspend, it may load not to an
intended URL, but to about:blank instead. Few mounths into
the work this bug caught my attention, and after dozens of
unsuspends I've come with this:

    Immediately after tab switch, TGS starts a checker to
check and match suspended tabs with the extention state.
But it forks not immediately, and in a short time window
it's possible to unsuspend unchecked tab. It leads to the
following log: 'Resuspending unresponsive suspended tab'
and the call of `tabsReload(tab.id);`. On Chromium it's
perfectly fine, but in FF it just breaks everything
(posibly due to a bad code inside the browser engine) and
leads to loading blank page insted of a desired.

    Patch adds a flag so checker won't reload the tab and
let it unsuspend properly. This flag later will be cleared
by existing code in `handleUnsuspendedTabStateChanged`.